refactor: extract import cli module
